+++ History.md
@@ -54,6 +54,8 @@
   the client, don't cache the return value of `cursor.count()` (consistently
   with the server behavior). `cursor.rewind()` is now a no-op. #2114
 
+* Allow externally applied CSS style attributes to interop with Blaze dynamic style attributes.
+
 * Upgraded dependencies:
   - node: 0.10.28 (from 0.10.26)
   - uglify-js: 2.4.13 (from 2.4.7)